tests/submission/PhilipTaylor/canvas/2d.drawImage.animated.gif.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.drawImage.animated.gif</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.drawImage.html">drawImage</a>.<a href="index.2d.drawImage.animated.html">animated</a>.gif</h1>
<p class="desc">drawImage() of an animated GIF draws the first frame</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.drawImage.animated.image">2d.drawImage.animated.image</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("drawImage() of an animated GIF draws the first frame");
_addTest(function(canvas, ctx) {

deferTest();
setTimeout(wrapFunction(function () {
    ctx.drawImage(document.getElementById('anim-gr.gif'), 0, 0);
    _assertPixelApprox(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 2);
}), 500);


});
</script>
<img class="resource" id="anim-gr.gif" src="../images/anim-gr.gif">

<div id="log"></div>
</body></html>
